Fixed Magnitude, Expt, and Compare
Compare was only returning true if the quaternions were exactly the
same, with the same represantation( 1/2 0 0 0) (0.5 0 0 0) would return
false) (fixed)
Magnitude and Expt was waiting for wrong input. ANd Magnitude was giving
a number as output, I changed the output to a quartenion with 0i 0j 0k,
just formalizing the output so if we have to deal with it in the ui when
showing o the user we can.